gh_214 M15-D: structure splits + comment diet (zero behavior change)
- e2e-checkpoint-exec.ts: restoreCheckpoint/bakeStageCheckpoint (+
  FlowExecError, re-exported) split out of the 985-line orchestrator;
  type-only import cycle, byte-identical behavior.
- up.ts: the sandbox/workspace launch prune (duplicated between dry-run
  and native run) is ONE sandboxDropSet helper — the two copies can no
  longer drift.
- Comment diet (skelly-approved, comments-only, agent-verified): the
  seam-getter pattern documented ONCE in base-command's header with 1-3
  line getter docs (unique contracts kept: stateDir arg, MAJOR-D,
  BLOCKER-B, applyInstanceEnv ordering); up.sh line-number citations
  dropped across prep/profiles/reset/migrate/base-command while keeping
  every behavioral rationale and bare up.sh provenance names.
- describeResolved module split + assembler dedup deliberately NOT
  taken (plan Q2: defer to the next feature touching both).

868 tests green; five captured CLI outputs byte-identical to the
pre-M15 baseline; lint warnings 83 -> 74.
